# documents/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "User API",
        "description": "API para gestionar usuarios en el sistema",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/court": {
            "post": {
                "tags": [
                    "Court"
                ],
                "summary": "Crear una pista deportiva",
                "operationId": "b0ae1a074698e1a83284a98ebdedfc34",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Court A"
                                    },
                                    "sport_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Court created"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/court/{id}": {
            "get": {
                "tags": [
                    "Court"
                ],
                "summary": "Obtener detalles de una pista",
                "operationId": "96e02c1be2e3bb16647c15b84b042119",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles de la pista",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Court login"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "name": "Court A",
                                                "sport": "1"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "No se ha encontrado este id"
                    }
                }
            },
            "put": {
                "tags": [
                    "Court"
                ],
                "summary": "Actualizar una pista deportiva",
                "operationId": "5edf6c5afb5280b660120596e7e7f747",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Court B"
                                    },
                                    "sport_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pista actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Court updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al actualizar la pista"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Court"
                ],
                "summary": "Eliminar una pista deportiva",
                "operationId": "aa789da1e82b99da0bb8ddd5c661cad6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la pista",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pista eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "string",
                                            "example": "Court deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/court/free": {
            "get": {
                "tags": [
                    "Court"
                ],
                "summary": "Obtener pistas disponibles",
                "operationId": "3adc713060b26479497543f5af3f46d0",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Fecha para consultar la disponibilidad",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-12-07"
                        }
                    },
                    {
                        "name": "member_id",
                        "in": "query",
                        "description": "ID del socio que consulta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "sport_id",
                        "in": "query",
                        "description": "ID del deporte",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pistas disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Court A"
                                                    },
                                                    "sport_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-07T17:26:38.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-12-07T17:26:38.000000Z"
                                                    },
                                                    "availableHours": {
                                                        "type": "array",
                                                        "items": {
                                                            "type": "string",
                                                            "example": "08:00:00"
                                                        }
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error al obtener disponibilidad",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Invalid parameters or no available courts"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}